Add files via upload
This program gathers customer lead information from Yellow Pages and prints it to a Word document.

I worked at a startup over the summer which provided a service to law enforcement. This program allows for you to look up department information based on the given variables of state, counties, and departments. This program searches Yellow Pages website and scrapes the necessary information from online and prints it to a word document. 

For this program to work for you, you will need to change the variables of state, counties, and departments to what you want them to be. You will also need to change the file path name of the Word document that you would like to print to. This should be all you need to change in order to use this program.